delivery_carrier_preference: Extract test setup in common class
